Refactor directory #17
new app name
new folder structur
