+++ src/assets/classlists/classlist_2024_08.txt
