+++ src/assets/classlists/classlist_2025_08.txt
