+++ src/assets/classlists/classlist_2026_08.txt
